Change merge logic in AdaptiveRebinner
Instead of defining the direction of merging based on the neighbour bin
with the largest uncertainty, check whole candidate regions to be merged
and choose one that contains the smallest effective number of events.
With the previous logic, the algorithm tended to merge angular bins
together, while it is more reasonable to merge bins in the tails of the
mass distribution.